Turn off strict mode
We were getting linting errors because the new ES6 block bindings aren't
compatible with strict mode, so we can't use both.
I'm not a big believer in strict mode anyway, and if we have to choose,
I'd much rather use ES6 features.

So I just turned off strict mode for everything for now.