Add new appState module
This module will act as a central dispatcher for all events on the browser
side. When state needs to be changed, functions will be called on appState
and any updates applied will be fed back in at the App.jsx level.

In this way, all state is centralized and processed by the dispatcher, and
components are totally driven by props, and will not have to maintain any
state.