add a new module to observe the apps current connection state
This proved difficult to test, and may need to be tweaked or enhanced over
time. It is unclear to me exactly when we would want to show the overlay.

It seems to me that we should be watching for events inside the
mattermost window that will tell us when the webview is connected to
the mattermost server, as opposed to the current code, which seems to just
be watching the Electron window's connection back to the main process.