Wire up app state object to our components
We set up the appState object to emit a stream of events when the app state
is updated.

The App component listens to this stream, and rerenders when there are
state changes.

The Overlay has also been moved inside the TeamWebview, so each team can
eventually have its own connection status.