add new watch tasks
The new start-watch task starts the app and watches the main-process and
browser-side code, restarting the appropriate part of the app as
necessary. When the app exits, everything shuts down tidily.

This eliminates the other, messier config for the hot-loader, works better
in Electron, and is just plain easier to see and understand.